requirements.txt
fastapi
uvicorn
snowflake-connector-python
pydantic-settings
python-dotenv